backend/requirements.txt
# Core Dependencies
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
pydantic_core==2.33.1

# MongoDB
pymongo==4.12.0
mongomock==4.3.0

# MinIO
minio==7.2.15

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
argon2-cffi==23.1.0

# Environment & Utilities
python-dotenv==1.1.0
python-multipart==0.0.20
certifi==2025.1.31

# Testing
pytest==8.3.5
httpx==0.26.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
pytest-cov==4.1.0
